+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Code generation inflection logic didn't quote the last but first error at retry
+- Python generation prompt referred to SQL
+
 ### Changed
 
 - Installed [the datarobot-pulumi-utils library](https://github.com/datarobot-oss/datarobot-pulumi-utils) to incorporate majority of reused logic in the `infra.*` subpackages.
+- Snowflake prompt more robust to lower case table and column names
 
 ## [0.1.8] - 2025-03-27